+++ frontend/src/components/MyPage/containers/ReviewManageContainer.tsx
@@ -4,6 +4,7 @@ import ReviewManage from '../ReviewManage';
 import { ManageCardItemType } from '../../../typedef/MyPage/myPage.types';
 import { useRecoilState } from 'recoil';
 import { userState } from '../../../recoil/UserAtom';
+import Swal from 'sweetalert2';
 
 interface GetReviewList {
 	reviews: ManageCardItemType[];
@@ -17,8 +18,16 @@ const ReviewManageContainer = () => {
 	const [myReviews, setMyReviews] = useState<ManageCardItemType[]>([]);
 
 	const onClickDeleteBtn = (seq: number) => {
-		http.delete(`web/reviews/${seq}`).then(() => {
-			setMyReviews([]);
+		Swal.fire({
+			title: '리뷰 삭제',
+			text: '정말 삭제하시겠습니까?',
+			showCancelButton: true,
+		}).then((result) => {
+			if (result.isConfirmed) {
+				http.delete(`web/reviews/${seq}`).then(() => {
+					setMyReviews([]);
+				});
+			}
 		});
 	};
